+++ packages/server/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sonata-api/server
 
+## 1.0.15
+
+### Patch Changes
+
+- Fixes
+- Updated dependencies
+  - @sonata-api/http@1.0.17
+  - @sonata-api/access-control@1.0.14
+  - @sonata-api/api@1.0.31
+  - @sonata-api/common@0.1.17
+  - @sonata-api/entrypoint@0.0.9
+  - @sonata-api/server@1.0.15
+  - @sonata-api/system@1.0.19
+  - @sonata-api/types@0.1.30
+
 ## 1.0.14
 
 ### Patch Changes