+++ packages/sonata-build/CHANGELOG.md
@@ -1,5 +1,15 @@
 # sonata-build
 
+## 1.0.12
+
+### Patch Changes
+
+- Fixes
+- Updated dependencies
+  - @sonata-api/common@0.1.17
+  - @sonata-api/system@1.0.19
+  - @sonata-api/types@0.1.30
+
 ## 1.0.11
 
 ### Patch Changes